// src/app/model/manage-booking/Service/Service.ts
export class Service {
  id: number;
  serviceType: string;
  description: string;
  roomNumber: string;
  servicePrice: number;
  position: number;
  bookingId: number;
  shared: true;
  date: string;
  orderId: number;

  organisationId : number;
  imageUrl	: string;
  logoUrl	: string;
  name	: string;
  afterTaxAmount : number;
  beforeTaxAmount: number;
  taxAmount : number;
  taxPercentage : number;
  businessType	: string;
  count : number;
  propertyReservationNumber : string;
  paymentId: number;
  autoExtend: boolean;
  balanceAmount: number;
  paidAmount : number;
  discountAmount : number;
  discountPercentage : number;
    constructor() {}
  }
  